put roi calculation functions into util file
